Implemented size_t usage

+++ core/batch.cpp
@@ -33,14 +33,14 @@ namespace RayGene3D
 {
   Batch::Batch(const std::string& name,
     Config& config,
-    const std::pair<const Entity*, uint32_t>& entities,
-    const std::pair<const Sampler*, uint32_t>& samplers,
-    const std::pair<const std::shared_ptr<View>*, uint32_t>& ub_views,
-    const std::pair<const std::shared_ptr<View>*, uint32_t>& sb_views,
-    const std::pair<const std::shared_ptr<View>*, uint32_t>& ri_views,
-    const std::pair<const std::shared_ptr<View>*, uint32_t>& wi_views,
-    const std::pair<const std::shared_ptr<View>*, uint32_t>& rb_views,
-    const std::pair<const std::shared_ptr<View>*, uint32_t>& wb_views)
+    const std::pair<const Entity*, size_t>& entities,
+    const std::pair<const Sampler*, size_t>& samplers,
+    const std::pair<const std::shared_ptr<View>*, size_t>& ub_views,
+    const std::pair<const std::shared_ptr<View>*, size_t>& sb_views,
+    const std::pair<const std::shared_ptr<View>*, size_t>& ri_views,
+    const std::pair<const std::shared_ptr<View>*, size_t>& wi_views,
+    const std::pair<const std::shared_ptr<View>*, size_t>& rb_views,
+    const std::pair<const std::shared_ptr<View>*, size_t>& wb_views)
     : Usable(name)
     , config(config)
     , entities(entities.first, entities.first + entities.second)
