+++ core/d11/d11_pass.h
@@ -52,7 +52,7 @@ namespace RayGene3D
     const std::shared_ptr<Config>& CreateConfig(const std::string& name,
       const std::string& source,
       Config::Compilation compilation,
-      const std::pair<const std::pair<std::string, std::string>*, uint32_t>& defines,
+      const std::pair<const std::pair<std::string, std::string>*, size_t>& defines,
       const Config::IAState& ia_state,
       const Config::RCState& rc_state,
       const Config::DSState& ds_state,
@@ -67,9 +67,9 @@ namespace RayGene3D
       Pass::Type type,
       uint32_t size_x,
       uint32_t size_y,
-      uint32_t layers,
-      const std::pair<const Pass::RTAttachment*, uint32_t>& rt_attachments,
-      const std::pair<const Pass::DSAttachment*, uint32_t>& ds_attachments);
+      size_t layers,
+      const std::pair<const Pass::RTAttachment*, size_t>& rt_attachments,
+      const std::pair<const Pass::DSAttachment*, size_t>& ds_attachments);
     virtual ~D11Pass();
   };
 }
