+++ core/vlk/vlk_view.h
@@ -59,8 +59,8 @@ namespace RayGene3D
     VLKView(const std::string& name,
       Resource& resource,
       Usage usage,
-      const View::Range& mipmaps_or_count = Range{ 0, uint32_t(-1) },
-      const View::Range& layers_or_stride = Range{ 0, uint32_t(-1) },
+      const View::Range& mipmaps_or_count = Range{ 0u, size_t(-1) },
+      const View::Range& layers_or_stride = Range{ 0u, size_t(-1) },
       View::Bind bind = View::BIND_UNKNOWN);
     virtual ~VLKView();
   };
